Make Python direct references and Git URLs work
Direct references are defined in PEP 440, and basically (as far as this
code cares) just put the name of the package at the beginning of the
line, followed by an '@' sign, then the URL (as if it were a bare URL).

Git URLs work by either cloning or fetching, then checking out the
specified committish.